Fixed the white border rendering of captain dash, by casting the position to int when drawing.

+++ Dash/GameObjects/GameObject.cs
@@ -61,7 +61,7 @@ namespace Dash
         {
             if (sprite != null)
             {
-                dc.DrawImage(sprite, position.X, position.Y, sprite.Width, sprite.Height);
+                dc.DrawImage(sprite, (int)position.X, (int)position.Y, sprite.Width, sprite.Height);
             }
 #if DEBUG
             foreach (Rect collisionBox in collisionBoxes)
